chore: isolate user notification service
