+++ backend/routes/api.js
@@ -2,7 +2,8 @@ const express = require('express');
 const router = express.Router();
 const path = require('path');
 const fs = require('fs').promises;
-const { getAllServices, createUserNotificationFromLog } = require('../services/fetchServices');
+const { getAllServices } = require('../services/fetchServices');
+const { createUserNotificationFromLog } = require('../services/createUserNotificationFromLog');
 const serviceControlRouter = require('./serviceControl');
 
 router.use('/service-control', serviceControlRouter);
